+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Clear Filters and Filter Sites buttons moved back from filter to map component
 -   Table styling for results and result modal
 -   If no location description for sensors in table, display N/A
+-   Disallowed closing of the login dialog
+-   Redirecting to landing page on logout
 
 ### Fixed